Refactor ethers-prometheus in preparation
